Добавил получение продуктов из базы данных

+++ Models/ProductViewModel.cs
@@ -8,7 +8,7 @@ namespace DeamonSharps.Shop.Simple.Models
     public class ProductViewModel
     {
         public string Name { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int ProductId { get; set; }
         public ProductViewModel GetDefaultSet()
         {
@@ -16,7 +16,7 @@ namespace DeamonSharps.Shop.Simple.Models
             return new ProductViewModel()
             {
                 Name = "Матан",
-                Price = 199.99,
+                Price = 199.99M,
                 ProductId = 0
             };
         }
